refactor; add resolveAssetPathsByRouteId
